QStringView: optimize search functions called with a single char
Most of these functions have inline versions of the QChar overload, so
we call them if the size of the input QStringView is known to be 1 at
compile time. For those that don't have an inline version (count(),
lastIndexOf()), this change simplifies the call sequence by passing a
single QChar or char16_t instead of a full View object.

Change-Id: I75cca6a0c3c4664cea77fffdebe0b798e7443537
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>
Reviewed-by: Ahmad Samir <a.samirh78@gmail.com>