QStringConverter: allow appendToBuffer() to write to the buffer
Even if we later decide that the data we wrote was incomplete. This
allows us to use the buffer as a scratch area, improving performance. I
need this for a faster SIMD encoding, because it stores a tentative
conversion to US-ASCII before searching where the last valid conversion
was. I was getting:

   Actual   (buffer)    : "\x00""dcdcdcd"
   Expected ("cdcdcdcd"): cdcdcdcd

Amends the tests added by f1c0bd2e06fdee87b255d790a42a361f3433cf24 for
finalize().

[ChangeLog][Important Behavior Change] The appendToBuffer() functions
in QStringEncoder and QStringDecoder may write to parts of the
provided output beyond the returned pointer. Caller code must not
assume the data beyond that remains unmodified.

Pick-to: 6.10 6.9 6.8
Change-Id: If59c4017cac9ee82ca45fffd29cceb9832ae98e5
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>