QTextBoundaryFinder: fix missing initialization of 'unused'
Coverity complained about a missing initialization of the 'unused'
field.

While we don't seem to have any inline API that could possibly depend
on the value of the field, not initializing reserved and unused
members has in the past caused problems for later extensions, so do it
now.

I don't expect problems here, since all SMFs are out-of-line, but
better safe than sorry.

Amends the start of the public history.

Coverity-Id: 11716
Coverity-Id: 11717
Coverity-Id: 11718
Coverity-Id: 444240
Pick-to: 6.10 6.9 6.8
Change-Id: Id25d9bb77fd0f03cdfa8dd809cfe662263b727a4
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>