QString: make qt_section_chunk a plain aggregate
The type doesn't need to have constructors, certainly not a default
constructor that doesn't initialize the 'length' member (so that even
value-initialization leaves it uninit'ed).

Since we don't depend on C++20, yet, this requires the use of {}
instead of () for initialization.

As a drive-by, fix the errornous spaces inside QStringView{ *this }.

Lack of initialization pointed out by Coverity.

Amends d807a16e7f9555e02991fc86ac5085ebe0101ef3.

Coverity-Id: 54563
Pick-to: 6.10 6.9 6.8
Change-Id: I8111ab253934a74e5f466a0457a9f7f3fb1f2721
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>