QObject: disconnect(): optimize finding the index of `method`
If `method` is wrapped in the SLOT() macro, we can skip searching
signals by using indexOfSlot().

Note that the string-based connect() overload uses the same technique.

In disconnect() there is one more usage of indexOfMethod(), but the
whole `try` block containing it will be refactored in a subsequent
commit.

Use a named-lambda, because it'll be re-used in later commits.

Task-number: QTBUG-139128
Change-Id: I7ac8417bf90e7f15c121163e4167af5ec2e00c9f
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>