a11y iOS: Announce changes to name/description to VoiceOver
Unlike macOS AppKit iOS UIKit does not offer an API to notify
VoiceOver about a value/title change. Also there is no public
API to query the system for the element which is currently
focused.

Both these caveats are addressed in this commit: the element with
active a11y focus is tracked; if the Accessible.name or
Accessible.description for *this* element changes, VoiceOver
is informed via UIAccessibilityLayoutChangedNotification and
refocuses the updated element.

Task-number: QTBUG-139275
Pick-to: 6.8 6.9 6.10
Change-Id: Ief1a2928aac138fe3f7afd9571838ed7713b327e
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>