QByteArray/QString: fix a few unintended copies
In reallocData() and reallocGrowData(), when we create a completely
new QArrayDataPointer, we merely copy-assigned it to this->d in the
end. This needlessly ups and downs the atomic ref-count, so use a
swap() instead.

Found by Coverity, which wants us to use std::move(), but I believe
the QtCore maintainer would prefer the swap here, so use "C++98 move".

Amends
- 20041afe3bdb3d3476c02897b8da288cba6011f4 QByteArray::reallocData()
- 287ace562ee5ddff22f7dbf4e49ae5f0520f2308 QString::reallocData()
- 4b9ec005340ac0d8aa103009c0f8ceda623de8cc QByteArray::reallocGrowData()
- a2b1b0292d3d9909a690ba701f0a447cdf1a66fa QString::reallocGrowData()

Pick-to: 6.10 6.9 6.8
Coverity-Id: 425077
Coverity-Id: 425941
Coverity-Id: 426231
Coverity-Id: 427184
Change-Id: I037f38c6541068b91b55343a20f149a772ba79c0
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>