CMake: Automatically generate a vcpkg manifest
[ChangeLog][configure] Added a -generate-vcpkg-manifest configure
option. This generates a vcpkg.json file in the build directory. To just
generate the manifest without configuring Qt, run configure with
-generate-vcpkg-manifest -dry-run.

[ChangeLog][configure] When configuring with -vcpkg, generate a vcpkg
manifest file in the build directory, and set VCPKG_MANIFEST_DIR to the
build directory to use that manifest file. You can turn off the manifest
generation with -no-generate-vcpkg-manifest. You can turn off setting
the manifest directory by passing -- -UVCPKG_MANIFEST_DIR to configure.

Configuring with pure CMake won't generate a vcpkg.json.

This patch adds a low-level CMake API for creating vcpkg manifests in
QtVcpkgManifestHelpers.cmake.

The functions qt_find_package and qt_feature gained arguments for
creating dependencies in the manifest and add them to features.

This adds a package dependency (not guarded by any feature):
    qt_find_package(WrapSystemFreetype 2.2.0 MODULE ...
        VCPKG_PORT freetype
        VCPKG_VERSION 2.3.4
    )

This adds a package dependency `freetype` to the feature `freetype`:
    qt_find_package(WrapSystemFreetype 2.2.0 MODULE ...
        VCPKG_PORT freetype
        VCPKG_ADD_TO_FEATURE freetype
    )

This adds a feature `freetype` with `gui` as its "parent dependency":
    qt_vcpkg_scope(gui)
    ...
    qt_feature("freetype" PUBLIC PRIVATE
        SECTION "Fonts"
        LABEL "FreeType"
        PURPOSE "Supports the FreeType 2 font engine (and its supported font formats)."
    )

If features are turned on/off via configure arguments, we set the CMake
variables VCPKG_MANIFEST_NO_DEFAULT to ON and VCPKG_MANIFEST_FEATURES to
the list of corresponding vcpkg features to only install libraries that
are needed for the build.

For a per-repo build, we create a vcpkg.json with dependencies only for
that repository.

For a top-level build, we create a vcpkg.json with all dependencies of
all repositories, excluding those that have been skipped (with
configure's -skip argument).

Task-number: QTBUG-73760
Change-Id: I6cc449d712f32d9e564acecaede5a7b31f309a8c
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>