Doc: Make 'Platform Notes' optional in template
Currently the app-examples-template.qdoc file contains a mandatory
'Platform Notes' section. This section lists the different platforms
that the example can be run on. Since all Qt application examples are
cross-platform, this information is not necessary.

This patch states that the 'Platform Notes' section is optional and
should be used if there are any plaform limitations or exceptions.

Fixes: QTBUG-139286
Pick-to: 6.10
Change-Id: Iad10b1cc5a5ca4880e92327d1ffd77b61988e72d
Reviewed-by: Andreas Eliasson <andreas.eliasson@qt.io>