Use QUtf8Functions::nextUcs4FromUtf8() in the JSON parser
Can't say this makes the function much simpler, and I didn't find a
Coverity issue about this use, but be consistent and use the new
function here, too. Besides, if we port Parser to qchar8_t, then it
will be simpler, and we might be able to remove the scanUtf8 function
completely.

Pick-to: 6.10 6.9 6.8
Change-Id: Ib6fa7aac14b942b4002163dccc77513a71378ea1
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>