Teach QImage::toCGImage() about most of our image formats
Instead of hard-coding the mapping between QImage formats and the
corresponding CGBitmapInfo we now build up the individual bits
of the bitmap info based on the pixel format of the image.

This gives us support for formats like RGB30 and the various
float formats that are important for HDR.

The qt_mac_cgImageFormatForImage helper now also resolves the
correct bits per component, as well as the color space, so
that consumers can't make any mistakes in what it passes to
the CoreGraphics APIs.

The validity of the format mappings has been tested by
saving both the QImage and the CGImage to a PNG file,
and comparing the results visually.

Once we have the same support for converting CGImageRefs
to QImage we can do this round-trip-verification in a test.

[ChangeLog][QtGui] QImage::toCGImage() now supports most
of the QImage formats. If a format is not supported the
function will return a null CGImageRef.

Change-Id: I3378ca00f7321e0c2c9d9a88a0dd25f5a893c56b
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>