QArrayData: Add MSVC attributes to MALLOCLIKE
__declspec(allocator) marks a function as something that allocates
memory, like malloc and new. This doesn't change optimizations in any
way, it is just for event tracking and the debugger.

__declspec(restrict) is used to indicate that any memory pointed to by
a pointer returned by the function is not aliased by any other pointer.
This is useful for the compiler to optimize memory access, namely it
can reorder stores and loads without having to take into account that
the memory might be accessed through another pointer.

Change-Id: Ic1320f3a4c17a12c9ac2c45eeb9ed70822380e6e
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>