Add QFileSelectorPrivate::removeStatics(const QStringList &)
There's currently no way to clear selectors that were added by external
modules.

For Qt Quick in particular, this means that QML modules that are loaded
by the type loader can add static selectors to the global list, and if
that module is later unloaded, those selectors can't be removed, and
will exist for as long as a Qt application is running.

It would make sense for selectors that were added in
QQmlExtensionPlugin::registerTypes() reimplemented functions to be removed
from the list when QQmlExtensionPlugin::unregisterTypes(), otherwise it
would count as littering.

Change-Id: I5dd66a175b3f7eecf793d67394cde136175e5d1f
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>