Teach qt_mac_toQImage about most of our image formats
Instead of always converting the image to Format_ARGB32_Premultiplied
potentially flattening formats with higher bit depths, we now try
to respect the original format of the CGImage.

This gives us support for formats like RGB30 and the various
float formats that are important for HDR.

A new qt_mac_imageFormatForCGImage helper has been added that
matches qt_mac_cgImageFormatForImage but in reverse.

During the conversion we try to adopt the existing data if
possible, which should be zero-copy despite the function name
CGDataProviderCopyData indicating otherwise.

[ChangeLog][Apple] Converting from CGImageRef, used for example
in the HEIC/HEIF image reader, now support a wider range of image
formats, including QImage::Format_RGB30 and the various floating
point formats.

Change-Id: Id49ee95f67cbe53819672ae52387917e0d7a61e1
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>