AndroidTestRunner: don't print crashes by default
In most cases after a crash, logs will be too long and will
flood the test output, making it not ideal and mainly unexpected.
Instead of printing the crash logs and logcat by default, have
it done only when it's explicitly requested via --show-logcat.

For CI though, add the --show-logcat by default so we don't lose
that verbosity.

Pick-to: 6.10
Change-Id: I2306b97378b6218ed22ba98aa50deefa9c9cd968
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>