Android: fix deadlock protector locking in QAndroidEventDispatcher
The order of conditions in QAndroidEventDispatcher::processEvents()
causes excessive and unnecessary locking for AndroidDeadlockProtector
with every call to processEvents() and in the majority of the cases
it's not needed at all. QAndroidEventDispatcher is mainly used when
setting the app's state or when creating the window surface, for that
reason we end up with calls first locking the deadlock protector
and then check m_stopRequest that most of the time is false. To avoid
any issues with trying to acquire same lock over and over and to be
more effecient the check for m_stopRequest needs to comes first.

Pick-to: 6.5 6.8 6.9 6.10
Fixes: QTBUG-132695
Change-Id: Iaa9a019ad666e2da073454d6473bba65c6c18418
Reviewed-by: Ville Voutilainen <ville.voutilainen@qt.io>