q20::countr_zero: add support for the TZCNT instruction with MSVC
This intruction introduced by Intel with the Bit Manipulation
Instructions (BMI) with the Haswell generation and later adopted by AMD
when they also adopted AVX2. Like LZCNT, this instruction reports the
size of the input if it was zero.

The processor has 8- and 16-bit forms of this instruction, but MSVC has
no intrinsics for them, so we still need to check for zero input, in
which case we may as well emit the old BSF instruction, for which we get
the input-was-zero status from the instruction itself, saving on some
code emission.

Change-Id: I116169a2d8b6ae96936ffffd563084a3009f2dce
Reviewed-by: Ivan Solovev <ivan.solovev@qt.io>