zlib: don't re-define Byte
The preprocessor is evil: by re-defining Byte, we cause a clash with the
MacTypes.h typedef, causing build errors with unity builds. It may also
clash with other uses of using Byte as identifier such as in
qbytearrayview.h or qmimemagicrule_p.h.

We do not need to re-define any of the typedefs for vendoring zlib,
as they won't affect globally visible symbols

Pick-to: 6.5 6.8 6.9 6.10
Change-Id: I978d76659b09e54c2a1efdecbb9529560884824d
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>