QUnicodeTables: abstract access to Properties::cases
This is in preparation for storing this information in a separate
array to save space by removing the many duplicates in there.

Pick to all active branches to have the same internal API going
forward, even if we don't pick the storage optimization back as far.

Qt 6.5 doesn't have QSpan, yet (not even as private API), but
returning a reference to const std::array<.,4> will be an adequate
replacement. To enable that without casting, convert `cases` from a C
array to a std::array. For some reason, this requires extra
parentheses, so add them.

Task-number: QTBUG-139427
Pick-to: 6.10 6.9 6.8 6.5
Change-Id: I5331fd6d71a6a447b0445d8235b5eb8e38178e2e
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>