wasm: implement exclusive suspend mode
This mode supports suspending the instance while waiting
for a single event only. All other events are queued for
later delivery.

This is useful for async calls made behind synchronous
API, where we don't want to run general event processing
while the instance is waiting for the async call to
complete. This is for example the case for file read()
type calls.

Change-Id: I8a8e8b15fea7a60a6bf069812294447505bdc717
Reviewed-by: Morten Johan Sørvig <morten.sorvig@qt.io>
Reviewed-by: Even Oscar Andersen <even.oscar.andersen@qt.io>