QRM: fix build when a std::*map is used as the range
While unorthodox, a std::map is sequentially iterable, and yields
std::pair instances as rows. Those should then result in two columns.

However, rows in a map cannot be rotated, because swap is explicitly
deleted for the pair type (which uses a const key type). So test that we
can swap the dereferenced iterators (testing for std::rotate still
passes, so we can't test for that directly).

Pick-to: 6.10
Change-Id: I12f7c06cca9363bdb3d0497ddbbf101d05bd9e9f
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>