Fix snippet compilation in corelib
The snippets use deprecated APIs, so the compilation fails when
building with QT_DISABLE_DEPRECATED_UP_TO set to a recent Qt version.
Fix it by wrapping the relevant code in the proper #ifdefs.

In QCoreApplication snippet, updeted QMouseEvent to use the modern API.

Remove the foreach() examples from the QVariant docs, because the usage
of foreach() is discouraged.

Amends:
* c1889bd953892428a474822168c10a4f983fb3ae for QDateTime snippets
* 55dc3ddaa7523a3e470c98271ea55fc7d2f3c414 for qglobal snippets
* d9251a64367e15f234718d9b3ed88d2f7bcfd0c6 for the rest

Change-Id: Idba2c115e2efb66b4cec707296ddfc38de9f9d25
Reviewed-by: Oleksii Zbykovskyi <Oleksii.Zbykovskyi@qt.io>
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>