util/unicode: Extract Method parseHexList() from readSpecialCasing()
This function, too, is duplicated all over the place, so centralize
it. Also use modern parsing techniques, like QStringTokenizer instead
of split() and QVarLengthArray instead of QList, that make the
function much more efficient.

Use it in readCaseFolding(), too, which is straight-forward.

There are many more potential users of this function; I'll port
them one by one in follow-up patches, though most are reading hex
ranges, so I'll add a function for that next.

Amends the start of the public history.

Pick-to: 6.10 6.9 6.8 6.5
Change-Id: I545a22d65a3baeaa850a7d658dcf466d2284b0fa
Reviewed-by: Ahmad Samir <a.samirh78@gmail.com>