QRM test: reference the view from the delegate correctly
A delegate that's getting recycled might not have a parent, resulting
in debug output from the binding to "parent.width". The correct way for
the delegate to reference the view it's bound to is to use e.g.
"ListView.view". "list" would work as well if we'd set

pragma ComponentBehavior: Bound

but that's just cluttering the code here.

Pick-to: 6.10
Change-Id: Ic6935a238029aad0d235a50fa078d44d62224e6a
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>