QRM docs: clarify direct access to the range, add section on subclassing
Those two are related: if we really didn't allow direct access to the
range after QRangeModel has been constructed, then subclassing would be
practically impossible. We do allow it, as long as no view (or similar)
operates on the model.

Clarify that in the \note, and add a snippet showing how to subclass
QRangeModel to build an encapsulated class that provides type- and
structure-aware APIs.

Pick-to: 6.10
Change-Id: I43836c750e372fd55c63bb75d89109524c87206c
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>