Core: Optimize QObjectCompatProperty's bindingWrapper
Instead of running the binding with empty input and then painstakingly
trying to compare the result afterwards, we can just pass the previous
value into the binding function like we do with all other kinds of
properties. The binding function will return false if it determines the
values are equal.

Change-Id: Ic6ecfb8982dfad84645cf95918a95f5d25f049f0
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>