util/unicode: Extract Method parseHex()
This code is used all over the place, so put it into a convenience
function where we can arguably put in a bit of effort to optimize
things, to wit: use QByteArrayView, and provide a better error message
than inline code could afford. E.g. the LastValidCodePoint check was
previously only in readUnicodeData().

For use in readUnicodeData(), add lineNo tracking.

In readBidiLine(), we can now drop the replace(' ', ""), because
parseHex() already trims. As a consequence, the lambda can take the
QByteArray by cref now.

There are many more places where this could be used, but they
represent higher-level constructs for which I'll add higher-level
helper functions.

Amends the start of the public history.

Pick-to: 6.10 6.9 6.8 6.5
Change-Id: Ic8f59f6509da1a0deb47a46cfaf160abb20c067e
Reviewed-by: Ahmad Samir <a.samirh78@gmail.com>