Fix memory leak in QCoreTextFontDatabase::addApplicationFont
The return value of CTFontDescriptorCreateCopyWithAttributes already has
retain count 1, and CFArrayAppendValue retains it again - so it leaks.
Due to this, memory is leaked every time an application font is added
on macOS from a data buffer. Wrap the the value in a QCFType to ensure
the local reference gets released.

Fixes: QTBUG-139600
Pick-to: 6.8 6.9 6.10
Change-Id: Iaedfea977627d214fe229ee675f321ad8255775a
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>