Give QLocaleData a dataForLocaleIndex(qsizetype)
Adding so that internal tests can use it in combination with
findLocaleIndex() to get locale data by the same parameters as we use
for locale. It incidentally packages an assertions used in most places
we were doing lookups in the locale_data[] array, so deploy it there.
Adapt one place which was accessing the array without that assertion
to condition its iteration on the bounds.

Change-Id: Ib3159184c7d84af9b194731c152549410acfdd6d
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>