QImageWriter: Don't load handlers from plugins if we found a built-in one
We never hit this case before, because all of our image handlers
declared their supported formats statically, but we can potentially
end up in this situation for handlers that report support for formats
that they have not declared.

The delete of handler in the loop was added because Valgrind was
complaining -- not because we actually supported or wanted plugins
to override the built in handlers unless they explicitly declared
they supported the format.

Pick-to: 6.10 6.9 6.8
Change-Id: If38f76345f71d419598b7834c83bbd5d07f757e1
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@qt.io>