QRM: remove explicit cbegin/cend usage and related code
In const member functions of the implementation class, the argument to
our begin/end/pos wrappers will always be const; so we can rely on
begin/end/pos overloads for const arguments, and don't have to
explicitly call cbegin/cend/cpos.

This removes a bunch of wrapper code, and reduces the requirements for
valid ranges for those with begin/end overloads, making it easier to
support more types from std::views in the future.

Pick-to: 6.10
Change-Id: Id8447e5fcf2d37f78d09ce79e07796540fcd027e
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>