QRM: store ranges that don't provide begin/end on const as mutable
This makes it almost possible to use certain std::views, like filter or
drop_while, with QRangeModel. Only do that for such ranges, as otherwise
we'd detach from Qt's implicitly shared range types.

Almost, because we still assume that we can call begin/end on a const
range in our size() implementation. That needs a bit more work, and
probably only makes sense when we have a fetchMore implementation for
ranges that don't provide a constant-time size().

Nevertheless, add a filter-case to the auto and manual tests, but go
through std::ranges::subrange for now.

Change-Id: I07cd543d73746e6b47a03aceedee2365b6750321
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>