QRM: API review - make roleNames helpers static
They don't need to access the ImplBase type, and the fewer calls back
from qrangemodel.cpp code into the implementation header we have, the
lower the risk for future binary compatibility problems. By making the
helpers static, they can't call back, reducing that risk.

Pass the item model as a QAIM reference so that we can get the default
implementation.

Pick-to: 6.10
Change-Id: I18e3437e05c45abcee8efe3664e8f5d24d657a55
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>