QRM: factor the code that inserts elements out
That allows us to call the logic with a custom functor that populates
the new elements with data immediately.

For unclear reasons, the simple refactoring fails to build on MSVC, as
it seems to hink that row_type{} is an iterator that needs to be
dereferenced - which fails for the move-only row types like tree_row.
Pass nullptr instead.

Change-Id: Ic265f168ad07634282bd287f64bbd4e7e3493a26
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>