Android: subtract action bar height from top safe margins
Qt for Android apps by default disable the native action bar, however,
especially on older Android versions, it can happen that the safe
margins are being calculated around the time that the action bar
is being hidden. In such case we will end up with incorrect values.
To avoid that, when retrieving the top safe margin, check if that
value minus the action bar height is positive value, in which case
we can assume that the initial top value consists from a status bar
height + an action bar height. This has been noticed mainly in
Android 9 and 10.

Fixes: QTBUG-135808
Pick-to: 6.9 6.10
Change-Id: Iabaaf9a864bd3b13a5b672080ee1efdd5adf98fe
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>