uic/Python: Handle surrogates correctly
When a QString has surrogates (2 consecutive QChar's in UTF-16),
convert to UCS4 first.

Rewrite the formatStringSequence() helper (which operates
on sequences of char/QChar) and related helpers to operate
on uint instead of ushort and remove the non-type template parameter
specifying the encoding by introducing a characterCode() converter
helper to convert to uint.

Check the QString passed in for surrogates and convert to UCS-4
if they are found (using the 32bit \U escape marker for Python).

Pick-to: 6.10 6.9 6.8
Fixes: PYSIDE-3173
Change-Id: Ie1d1282d78ad80894cf67dd47ea6a332d7dfda25
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>