a11y: Let buddy widget know its label(s)
If a widget is set as the buddy widget of a QLabel,
that label is reported for the widget's QAccessible::Label
relation.

So far, in order to identify labels for which the
widget is set as the buddy, all sibling widgets
were inspected. There was already a comment mentioning
that all objects should ideally be checked, but that
would be too expensive.

Change the approach altogether: Instead of
traversing the widget tree, store all labels for
which the widget is set as the buddy in the widget
itself.

This is more efficient and ensures the relation
is reported correctly for non-siblings as well.

Fixes: QTBUG-139719
Change-Id: I6d517f8727cdde0de38b11d7c6906ecc55da1927
Reviewed-by: MohammadHossein Qanbari <mohammad.qanbari@qt.io>