QHttp2Connection GOAWAY overhaul
While trying to figure out what the problem may be in the linked bug
report, it became visible that the GOAWAY parts likely haven't been
exercised very well.

For one, we are supposed to ignore incoming streams beyond what we told
the peer was the last stream to be processed. But we also need to
process certain frames, like DATA and frames carrying HEADER content,
because they affect state for the entire connection, state that the
peer cannot roll back on their side.

There were also some bugs in our handling of a GOAWAY frame, such as
always assuming it was a client on the receiving end and, following on
from that, some issues with notifying streams that they have received
a GOAWAY.

Task-number: QTBUG-139692
Task-number: QTBUG-135049
Pick-to: 6.10 6.8
Change-Id: I0bc1f51787b7a9df81a6a9b855a390c8f8b30e5b
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>
Reviewed-by: Dennis Oberst <dennis.oberst@qt.io>