Introduce QIORing to abstract io_uring on Linux
Somewhat low level code, intended as a low level
abstraction.

The QIORing interface would also be used for Windows' IORing, developed
at the same time as the io_uring version.
There is some shared code and helper functions, but a lot of the code
in some way touch the platform specifics, so without yet more
abstractions quite some code is left as unique.

The fiveGiBReadWrite test case is currently EXPECT_FAIL because of the
MAX_RW_COUNT limit on Linux (and its inability to report >2GiB results).
Either we have to document this in the public-facing parts, or we need
to work around / iron over the issue. To be done in a follow-up patch.

We only ever notify the kernel of the work to be done after returning to
the event loop, unless the ring is full. Then we notify the kernel right
away in hopes it will manage to clear up some space to queue more.
We, ourselves, are not actually limited by the kernel ring buffers as
we keep a queue (in the form of a std::list) of pending tasks.

The reason why it's a std::list is that it guarantees stable references,
and that lets us use the pointer-to-task as the 'user_data' of the
submitted work, so we can easily access the task again when the kernel
adds it to the completion queue.

Task-number: QTBUG-136763
Change-Id: I9cb80a2b96a49f2a557bef3b0ad6d367d76f5ab8
Reviewed-by: Ivan Solovev <ivan.solovev@qt.io>