Implement a backend of QIORing for Windows
Using the IORing API on Windows to provide asynchronous I/O operations.
Some parts supplemented by using some blocking API where we don't have
an option (yet). Currently that just includes the 'stat'/size operation
as well as Open and Close. Though with Close we schedule a flush and
close the handle ourselves once the callback is invoked.

The API is quite limited so far, but sufficient for what we have now.
The implementation can be extended later as needed.

The Vectored I/O operations are not actually vectored unfortunately,
the Windows API requires page-aligned memory and sector-aligned file
offsets, which makes it really impractical to provide generically.

For a very limited time: limit the configure options to Windows 11.

Task-number: QTBUG-136763
Change-Id: Iee57a23358a71ab6bfd007ff15b760b65ea76406
Reviewed-by: Ivan Solovev <ivan.solovev@qt.io>