QDataStream: fix -Werror,-Wshorten-64-to-32 on 32-bit Clang
On 32-bit platforms qsizetype is int, so assigning a qint64 to it
narrows.

It's benign, because we perform the overflow check in the next line,
but that doesn't help -Werror.

Fix by making the narrowing explicit. As a drive-by, mark `n` const.

Amends fd48ce0b73c74dafd5db27bc1f2752ef665df7ef.

Reported-by: Karim Pinter <karim.pinter@qt.io>
Pick-to: 6.10 6.9 6.8
Change-Id: I14d95c54c501df14a0b76d11262f22605fbb583b
Reviewed-by: Øystein Heskestad <oystein.heskestad@qt.io>
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>