tst_QThread: further improve test on multiple threads waiting on wait()
Amends 2bce75a6b53cccbf9c813581b64eea87f3ab55fc which had itself amended
commit 5b5297fe87859f59a7aaf5e86a8915c00714fefa.

This commit adds a much lengthier explanation of the design of this test
-- which I had to write to understand and fix the problems anyway.

The fix itself consists of uses two QElapsedTimers in each thread, not
just one: one for the threads that time out and started before the
auxiliary threads, and one for threads that succeed and is started after
they have reached the synchronization point. Because the main thread
must now wait for the WaiterThreads to reach the synchronization point,
we needed to replace the startSema QSemaphore with a barrier instead.

Pick-to: 6.10 6.9
Fixes: QTBUG-139617
Change-Id: I5d86dd42f5cd9b87a55dfffd30fd47637ea3066a
Reviewed-by: Ivan Solovev <ivan.solovev@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>